chore: move e2e test to `tests/`
